+++ src/modules/customer/index.ts
@@ -1,2 +1,3 @@
 export * from './get-balance/get-balance';
+export * from './list-billing-history/list-billing-history';
 export * from './types';